:recycle: Improve src/handler
Removes unnecessary submoduling & adjusts some weird code
